scripts/transcription/requirements.txt
supabase
python-dotenv
requests
requests-toolbelt
anthropic
